backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyMySQL==1.1.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
